+++ services/notification/src/main/java/com/barlow/services/notification/worker/MessageProvider.java
@@ -1,0 +1,11 @@
+package com.barlow.services.notification.worker;
+
+import com.barlow.core.enumerate.DeviceOs;
+import com.barlow.services.notification.NotificationInfo;
+
+public interface MessageProvider {
+
+	NotificationMessage provide(String messageTitle, String messageBody, NotificationInfo.Subscriber subscriber);
+
+	DeviceOs supportedOs();
+}
